requirements.txt
-r base_requirements.txt
scipy==0.11.0
sphinx==1.1.3
django-sphinx-autodoc==0.2
gunicorn
wsgiref
ipython
supervisor==3.0
